requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
openai>=1.30.0
anthropic>=0.28.0
ddgs>=0.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
pydantic>=2.7.0
rich>=13.7.0
matplotlib>=3.8.0
pandas>=2.2.0
numpy>=1.26.0
